requirements.txt
altair==5.0.1
attrs==23.1.0
auth==0.5.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
dnspython==2.4.1
eventlet==0.33.3
falcon==3.1.1
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mongoengine==0.27.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.15.1
pymongo==4.4.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
vertexai==0.0.1
watchdog==3.0.0
Werkzeug==2.3.6
zipp==3.16.2
